+++ src/main/kotlin/it/czerwinski/intellij/wavefront/settings/ObjPreviewFileEditorSettingsState.kt
@@ -17,9 +17,11 @@
 package it.czerwinski.intellij.wavefront.settings
 
 import com.intellij.util.xmlb.annotations.Attribute
+import it.czerwinski.intellij.wavefront.editor.model.ShadingMethod
 import it.czerwinski.intellij.wavefront.editor.model.UpVector
 
 data class ObjPreviewFileEditorSettingsState(
+    @field:Attribute var defaultShadingMethod: ShadingMethod = ShadingMethod.DEFAULT,
     @field:Attribute var defaultUpVector: UpVector = UpVector.DEFAULT,
     @field:Attribute var showAxes: Boolean = DEFAULT_SHOW_AXES,
     @field:Attribute var axisLineWidth: Float = DEFAULT_AXIS_LINE_WIDTH,