+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]
 ### Added
 - Support for relative indices in OBJ files (#95)
+- Refresh action in 3D preview
 
 ### Changed
 - Use standard IDE error icons